LayoutTests/imported/w3c/web-platform-tests/navigation-api/per-entry-events/dispose-same-document-navigate-during-expected.txt


FAIL navigate() during a same-document-navigation-initiated dispose works (since it's after the previous navigation) assert_equals: expected "#2" but got "#1"
